backend/np-tools/requirements.txt
backoff==2.2.1
#bertopic==0.16.4
dask==2025.2.0
Flask==3.1.0
flask_restx==1.3.0
fuzzywuzzy==0.18.0
gensim==4.3.3
ipdb==0.13.13
langchain==0.3.20
langchain_community==0.3.19
langchain_openai==0.3.8
language_tool_python==2.9.0
llama_index==0.12.24
matplotlib==3.10.1
numpy
openai==1.66.3
pandas==2.2.3
pdfminer==20191125
pdfplumber==0.11.5
pyfiglet==1.0.2
pyLDAvis==3.4.1
PyPDF2==3.0.1
python-dotenv==1.0.1
regex==2024.11.6
Requests==2.32.3
scikit_learn==1.6.1
scipy
sentence_transformers==3.4.1
spacy==3.8.4
sparse_dot_topn==1.1.5
termcolor==2.5.0
tomotopy==0.13.0
tqdm==4.67.1
#umap==0.1.1
pyarrow
frontend==0.0.3
pymupdf
pdf2image
skops
faiss-cpu
dspy==2.6.3
bert_score
ollama
openpyxl